+++ intelmq/tests/lib/test_upgrades.py
@@ -13,6 +13,13 @@ V202 = {"test-collector": {
         "feed": "Feed"
     }
 },
+    "ripe-expert": {
+    "group": "Expert",
+    "module": "intelmq.bots.experts.ripe.expert",
+    "parameters": {
+        "query_ripe_stat_asn": True,
+    },
+}
 }
 V202_EXP = {"test-collector": {
     "group": "Collector",
@@ -21,7 +28,57 @@ V202_EXP = {"test-collector": {
         "name": "Feed"
     }
 },
+    "ripe-expert": {
+    "group": "Expert",
+    "module": "intelmq.bots.experts.ripe.expert",
+    "parameters": {
+        "query_ripe_stat_asn": True,
+        "query_ripe_stat_ip": True,
+    },
 }
+}
+
+DEP_110 = {"n6-collector": {
+    "group": "Collector",
+    "module": "intelmq.bots.collectors.n6.collector_stomp",
+    "parameters": {
+        "feed": "Feed",
+    },
+},
+    "cymru-full-bogons-parser": {
+    "group": "Parser",
+    "module": "intelmq.bots.parsers.cymru_full_bogons.parser",
+    "parameters": {
+    },
+},
+    "ripe-expert": {
+    "group": "Expert",
+    "module": "intelmq.bots.experts.ripencc_abuse_contact.expert",
+    "parameters": {
+        "query_ripe_stat": True,
+    },
+}
+}
+DEP_110_EXP = {"n6-collector": {
+    "group": "Collector",
+    "module": "intelmq.bots.collectors.stomp.collector",
+    "parameters": {
+        "name": "Feed",
+    },
+},
+    "cymru-full-bogons-parser": {
+    "group": "Parser",
+    "module": "intelmq.bots.parsers.cymru.parser_full_bogons",
+    "parameters": {
+    }},
+    "ripe-expert": {
+    "group": "Expert",
+    "module": "intelmq.bots.experts.ripe.expert",
+    "parameters": {
+        "query_ripe_stat_asn": True,
+        "query_ripe_stat_ip": True,
+    },
+}}
 
 
 def generate_function(function):
@@ -52,6 +109,12 @@ class TestUpgradeLib(unittest.TestCase):
         self.assertEqual(set(self.allfs), set(self.modulefs),
                          msg='v* functions in the module do not '
                              'match functions in __all__.')
+
+    def test_v110_deprecations(self):
+        """ Test v110_deprecations """
+        result = upgrades.v110_deprecations({}, DEP_110, False)
+        self.assertTrue(result[0])
+        self.assertEqual(DEP_110_EXP, result[2])
 
     def test_v202_fixes(self):
         """ Test v202_feed_name """